expert app test1

+++ FastAPI_SQL_assignment6/expert_app/app/crud.py
@@ -12,27 +12,33 @@ def get_students(db: Session, skip: int = 0, limit: int = 100):
     return db.query(models.Student).offset(skip).limit(limit).all()
 
 def create_student(db:Session, student: schemas.StudentCreate):
-    db_student = models.Student(firstname = student.firstname, lastname = student.lastname,
-                                 average = student.average, graduated = student.graduated)
+    db_student = models.Student(**student.model_dump())
     
     db.add(db_student)
     db.commit()
     db.refresh(db_student)
     return db_student
 
-def edit_student(db:Session, student: schemas.StudentEdit):
-    db_student = models.Student(**student.model_dump())
-    db.add(db_student)
-    db.commit()
-    db.refresh(db_student)
-    return db_student
-
-def remove_student(db:Session, student: schemas.StudentDelete, stu_id: int):
+def edit_student(db:Session, stu_id:int, student: schemas.StudentCreate):
     db_student = db.query(models.Student).filter(models.Student.id == stu_id).first()
-    db.delete(db_student)
-    db.commit()
-    db.refresh(db_student)
-    return {"message": "student with {stu_id} deleted"}
+    if db_student is None:
+        return False
+    else:
+        db_student = models.Student(**student.model_dump())
+        db.add(db_student)
+        db.commit()
+        db.refresh(db_student)
+        return db_student
+
+def remove_student(db:Session, stu_id: int):
+    db_student = db.query(models.Student).filter(models.Student.id == stu_id).first()
+    if db_student is None:
+        return False
+    else:
+        db.delete(db_student)
+        db.commit()
+        #db.refresh(db_student)
+        return True
 
 
 
@@ -42,14 +48,35 @@ def get_courses(db: Session, skip: int = 0, limit: int = 100):
     return db.query(models.Course).offset(skip).limit(limit).all()
 
 def create_student_course(db:Session, course: schemas.CourseCreate, stu_id: int):
-    db_course = models.Course(**course.model_dump(), owner_id = stu_id)
-    db.add(db_course)
-    db.commit()
-    db.refresh(db_course)
+    db_student = db.query(models.Student).filter(models.Student.id == stu_id).first()
+    if db_student is None:
+        return False
+    else:
+        db_course = models.Course(**course.model_dump(), owner_id = stu_id)
+        db.add(db_course)
+        db.commit()
+        db.refresh(db_course)
     return db_course
 
-def edit_student_course(db: Session, course: schemas.CourseEdit, stu_id: int):
-    pass
+def edit_student_course(db: Session, course: schemas.CourseCreate, stu_id: int, course_id: int):
+    db_student = db.query(models.Student).filter(models.Student.id == stu_id).first()
+    db_course = db.query(models.Course).filter(models.Course.id == course_id).first()
+    if (db_student is None) or (db_course is None):
+        return False
+    else:
+        db_course = models.Course(**course.model_dump())
+        db.add(db_course)
+        db.commit()
+        db.refresh(db_course)
+        return db_course
 
-def remove_student_course(db: Session, course: schemas.CourseDelete, stu_id:int):
-    pass
+def remove_student_course(db: Session, course_id: int, stu_id:int):
+    db_student = db.query(models.Student).filter(models.Student.id == stu_id).first()
+    db_course = db.query(models.Course).filter(models.Course.id == course_id).first()
+    if (db_student is None) or (db_course is None):
+        return False
+    else:
+        db.delete(db_course)
+        db.commit()
+        #db.refresh(db_course)
+        return True
